A lot of random changes (Literally random)
I should finally focus on single thing and not do 10 at the same time >.>
Added a lot of test run configurations
Added (not used for now) web browser dependency.
Bumped version to 2.0-DEVELOP
Moved time counting to the main class
Changed a bit prompts at the start
Added 403 Response Code handling (Still 500 and 203 handling to do, but they are super rare so)
Realised CF has a lot of weird stuff and my plan for better fallback is doomed for this project.
Improved warning handling in Manifest
Started work on the updater!
Added handling for 403 errors in the SyncManager