More work on the Updater + SettingsManager Saving Rework
Reworked SettingsManager saving capability, it now actually changes only lines it has to, instead of overriding the entire file!
Added to Settings Logger settings (finally)
Changed few arguments and variable names because of typos.
Some work on the Updater! (It already checks for the updates and shows you the GUI!!)
Updated Settings.json5